Windows Atomic Tests to TTP #7 (#131)
Summary:
Pull Request resolved: https://github.com/facebookincubator/ForgeArmory/pull/131

Converting atomics to ttps in Windows Atomic Red Team Tests
This ttp was 7/10 and it performs the follow function:
(Tor Proxy Usage - Windows)
This test is designed to launch the tor proxy service, which is what is utilized in the background by the Tor Browser and other applications with add-ons in order to provide onion routing functionality.
Upon successful execution, the tor proxy will be launched, run for 60 seconds, and then exit.

Reviewed By: godlovepenn

Differential Revision: D62394294

fbshipit-source-id: d7e0aff64134a7811f081ddf0d50acf6babebace