Windows Atomic Tests to TTP #7 Update
Summary:
Added error checking and some updates to Windows Atomic Red Team Tests #7

This ttp was 7/10 and it performs the follow function:
(Tor Proxy Usage - Windows)
This test is designed to launch the tor proxy service, which is what is utilized in the background by the Tor Browser and other applications with add-ons in order to provide onion routing functionality.
Upon successful execution, the tor proxy will be launched, run for 60 seconds, and then exit.

Differential Revision: D62876655