Windows Atomic Tests to TTP #7 Update (#138)
Summary:
Pull Request resolved: https://github.com/facebookincubator/ForgeArmory/pull/138

Added error checking and some updates to Windows Atomic Red Team Tests #7

This ttp was 7/10 and it performs the follow function:
(Tor Proxy Usage - Windows)
This test is designed to launch the tor proxy service, which is what is utilized in the background by the Tor Browser and other applications with add-ons in order to provide onion routing functionality.
Upon successful execution, the tor proxy will be launched, run for 60 seconds, and then exit.

Reviewed By: godlovepenn

Differential Revision: D62876655

fbshipit-source-id: db3746170e5750ae28566c7d6076a8ba1693b22d