Update pkill to continue script
when I run this TTP, it doesn't seem to run successfully whether or not Firefox is running. If Firefox is running, it closes but errors. When Firefox isn't running, it errors. I think it's because the `pkill` statement is returning status of 1. 

```
# firefox not running
% pkill -a -i "Firefox"
% echo $?
1
```

This PR adds `|| true` to tell the script to continue. I'm not sure why the second `pkill` is there but this unblocks this locally for me.